--- CONTRIBUTING.md
@@ -1,9 +1,0 @@
-#Contributing
-Hey user!
-If you found any issues please add a pull request or an issue!
-And if you want to add another mod feel free to add a pull request.
-I really want to invite you to create mods for Discord Bot Maker.
-
-If you have any questions contact me on Discord: Lasse#4890.
-Start contributing now!
-Lasse
