--- PULL_REQUEST_TEMPLATE.md
@@ -1,6 +1,0 @@
-Your Discord name with tag:
-Are you already a member of the DBM Mod Guild?
-Category of actions you added (Store Things/UI/...):
-
-
-If you aren't a  member of the Guild, make sure to join! Link is in the README file :kappa:
